requirements.txt
pywin32
pandas
colorama
browser-history
python-registry
pytest